requirements.txt
Flask
flask-debugtoolbar
Flask-WTF
Flask-RESTful
Flask-SQLAlchemy
Pillow
uwsgi
mysqlclient
psycopg2